requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-msi
azure-mgmt-subscription
requests